+++ packages/tokens/src/index.ts
@@ -2,6 +2,9 @@ export * from "./font-sizes";
 export * from "./font-weights";
 export * from "./fonts";
 export * from "./line-heights";
+export * from "./shadows";
+export * from "./transitions";
+export * from "./zIndices";
 
 export * from "./radii";
 export * from "./space";