+++ packages/tokens/src/transitions.ts
@@ -1,0 +1,3 @@
+export const transitions = {
+  basic: "cubic-bezier(0.25, 0.46, 0.45, 0.94) 0s",
+};