+++ packages/tokens/src/zIndices.ts
@@ -1,0 +1,3 @@
+export const zIndices = {
+  absolute: 999999,
+};